Revert "[fix] max_pool1d: composite compliance (#70900)" (#71992)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71992

This reverts commit ff9352e3977a4dc293c4a329173b726be7f8b1d1.

We are conservatively reverting this because it broke a test in functorch.
The original PR added a `_max_pool1d_cpu` operator. I'm not sure if it
is actually safe to revert this due to the addition of the new operator
(someone may have serialized it between now and then) but because it has
only been two weeks this should be fine.

Test Plan: - wait for tests

Reviewed By: jbschlosser, VitalyFedyunin

Differential Revision: D33882918

Pulled By: zou3519

fbshipit-source-id: f146e82e6b46690376b3d8825dc7f7da62e2c7de
(cherry picked from commit 1606333e6ce23d618863a9b0e504352bd55569bc)