[Quant][devs] Changed empty_quantized call for quantized tensor to resize_output (#71899)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71899

This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR removes the call to empty_quantized for quantized tensors and substitutes
it for resize_output, which works for quantized tensors, based on current understanding.
Using the dispatcher for this function was determined to be not practical as it would entail
a significant amoutn of duplicate code

Differential Revision:
D33809138
D33809138

Test Plan: Imported from OSS

Reviewed By: jerryzh168

Pulled By: dzdang

fbshipit-source-id: 5bacea37356547ceacea4b3f6b0141ac3a223dcf
(cherry picked from commit 3bb82ff3040c9a7905a3cfe8a57c69cfe0721955)