move version_h to shared build structure in Buck (#75964)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75964

This is already in the shared build structure for Bazel, but we need
to implement genrule for fbcode.

There's an xplat target that can't build in fbcode yet because the
dependencies don't line up, so we have to add a tag to exclude it.

ghstack-source-id: 154696020

Test Plan: Rely on CI

Reviewed By: malfet

Differential Revision: D35443900

fbshipit-source-id: 0768b29906c8218d7aebfdc7c18d69f59a0c9384
(cherry picked from commit bff47be441bd142392a07aa177be02e18aa86f1c)